pizarra, estadisticas, y mejoras

+++ conductor/formato.php
@@ -22,51 +22,41 @@ echo '      </td>';
 echo '  </tr>';
 echo '  <tr>';
 echo '      <td>';
-/*echo '      </td>';
-echo '  </tr>';
-echo '  <tr>';
-echo '      <td>';*/
 echo '          <h2 class="text-center">Empresa Nacional de Transporte</h2>';
+echo '          <h4 class="text-center"><i>Ficha Técnica de Conductor</i></h4>';
 echo '      </td>';
 echo '  </tr>';
 echo '';
-echo '</table><br><br><br><br><br>';
+echo '</table><br><br><br><br>';
 
 
     include_once("../config/init.php");
 
     $id = $_GET['id'];
 
-    $resultado_listado = Chuto::listar_chuto_byid($id);
+    $resultado_listado = Conductor::listar_conductor_byid($id);
 
     if ($resultado_listado == true) {
 
         foreach ($resultado_listado as $row) {
 
-    // echo "<div class='table-responsive'>";
         echo "<table align='center' class='table'>";
             echo "<thead>";
 
             echo "<tr>
                 <th class='text-center'>#</th>
-                <th class='text-center'>Placa</th>
-                <th class='text-center'>Placa Nueva</th>
-                <th class='text-center'>Serial de Carroceria</th>
-                <th class='text-center'>Serial de Motor</th>
-                <th class='text-center'>Marca</th>
-                <th class='text-center'>Tipo</th>
+                <th class='text-center'>Cedula</th>
+                <th class='text-center'>Nombre</th>
+                <th class='text-center'>Apellido</th>
                 </tr>";
             echo "</thead>";
 
             echo "<tbody>";
                 echo "<tr>";
-                echo '<th class="text-center" scope="row">' . $row->id_chuto . '</th>';
-                echo "<td class='text-center'>" . $row->placa_chuto . "</td>";
-                echo "<td class='text-center'>" . $row->placa_nueva_chuto . "</td>";
-                echo "<td class='text-center'>" . $row->serial_carroceria_chuto . "</td>";
-                echo "<td class='text-center'>" . $row->serial_motor_chuto . "</td>";
-                echo "<td class='text-center'>" . $row->marca_chuto . "</td>";
-                echo "<td class='text-center'>" . $row->tipo_chuto . "</td>";
+                echo '<th class="text-center" scope="row">' . $row->id_conductor . '</th>';
+                echo "<td class='text-center'>" . $row->cedula_conductor . "</td>";
+                echo "<td class='text-center'>" . $row->nombre_conductor . "</td>";
+                echo "<td class='text-center'>" . $row->apellido_conductor . "</td>";
                 echo "</tr>";
             echo "</tbody>";
         echo "</table><br><br>";
@@ -74,68 +64,76 @@ echo '</table><br><br><br><br><br>';
         echo "<table align='center' class='table'>";
             echo "<thead>";
             echo "<tr>
-                <th class='text-center'>Modelo</th>
-                <th class='text-center'>Año</th>
-                <th class='text-center'>Color</th>
-                <th class='text-center'>Observacion</th>
-                <th class='text-center'>Fecha Modificacion</th>
+                <th class='text-center'>Fecha de Certificacion</th>
                 <th class='text-center'>Sede</th>
-                <th class='text-center'>Estatus</th>
+                <th class='text-center'>Estado</th>
+                <th class='text-center'>Especificacion</th>
                 </tr>";
             echo "</thead>";
 
             echo "<tbody>";
                 echo "<tr>";
-                echo "<td class='text-center'>" . $row->modelo_chuto . "</td>";
-                echo "<td class='text-center'>" . $row->a_o_chuto . "</td>";
-                echo '<td class="text-center borde" style="background-color:' . $row->color_chuto_1 . ';"><b>' . $row->nombre_color_chuto . '</b></td>';
-                echo "<td class='text-center'>" . $row->observacion_chuto_estado . "</td>";
-                echo "<td class='text-center'>" . $row->fecha_chuto_estado . "</td>";
+                echo "<td class='text-center'>" . $row->fecha_conductor . "</td>";
                 echo "<td class='text-center'>" . $row->nombre_sede . "</td>";
-                echo "<td class='text-center'>" . $row->chuto_estado . "</td>";
+                echo "<td class='text-center'>" . $row->nombre_estado . "</td>";
+                echo "<td class='text-center'>" . $row->nombre_especificacion . "</td>";
                 echo "</tr>";
             echo "</tbody>";
 
         echo "</table><br><br><br>";
 
-        echo '<table style="border: solid 1px #D51C28">';
+        echo '<table align="center" style="border: solid 1px #D51C28">';
         echo '  <tr>';
-        echo '      <td>';
-        echo '          <h4>Foto de la Placa</h4>';
-        echo '          <hr>';
-        echo '          <img src="../img/chuto/placa/' . $row->id_chuto . '_placa.png" alt="Logo" width=350 />';
-        echo '      </td>';
-    /*    echo '  </tr>';
-        echo '</table>';
 
-            echo '<table style="border: solid 1px #D51C28">';
-        echo '  <tr>';*/
+        if(file_exists('../img/conductor/carnet/' . $row->id_conductor . '_carnet.png')) {
+
         echo '      <td>';
-        echo '          <h4>Foto del Serial de Carroceria</h4>';
+        echo '          <h4>Foto del Carnet</h4>';
         echo '          <hr>';
-        echo '          <img src="../img/chuto/serial_carroceria/' . $row->id_chuto . '_serial_carroceria.png" alt="Logo" width=350 />';
+        echo '          <img src="../img/conductor/carnet/' . $row->id_conductor . '_carnet.png" alt="Logo" width=300 />';
         echo '      </td>';
+        
+        }
+        
+        if(file_exists('../img/conductor/cedula/' . $row->id_conductor . '_cedula.png')) {
+
+        echo '      <td>';
+        echo '          <h4>Foto de la Cedula de Identidad</h4>';
+        echo '          <hr>';
+        echo '          <img src="../img/conductor/cedula/' . $row->id_conductor . '_cedula.png" alt="Logo" width=300 />';
+        echo '      </td>';
+        
+        }
+
         echo '  </tr>';
         echo '</table>';
 
-            echo '<table style="border: solid 1px #D51C28">';
+        echo '<table align="center" style="border: solid 1px #D51C28">';
         echo '  <tr>';
+
+        if(file_exists('../img/conductor/certificado_medico/' . $row->id_conductor . '_certificado_medico.png')) {
+
         echo '      <td>';
-        echo '          <h4>Foto del Serial de Motor</h4>';
+        echo '          <h4>Foto del Certificado Medico</h4>';
         echo '          <hr>';
-        echo '          <img src="../img/chuto/serial_motor/' . $row->id_chuto . '_serial_motor.png" alt="Logo" width=350 />';
+        echo '          <img src="../img/conductor/certificado_medico/' . $row->id_conductor . '_certificado_medico.png" alt="Logo" width=300 />';
         echo '      </td>';
+        
+        }
+
+        if(file_exists('../img/conductor/licencia/' . $row->id_conductor . '_licencia.png')) {
+
+        echo '      <td>';
+        echo '          <h4>Foto de la Licencia</h4>';
+        echo '          <hr>';
+        echo '          <img src="../img/conductor/licencia/' . $row->id_conductor . '_licencia.png" alt="Logo" width=300 />';
+        echo '      </td>';
+        
+        }
+
         echo '  </tr>';
         echo '</table>';
 
-        // echo '<table style="border: solid 1px #D51C28">';
-        // echo '  <tr>';
-        // echo '      <td>';
-        // echo '          <h4>Titulo</h4>';
-        // echo '          <img src="../img/chuto/seguro/' . $row->id_chuto . '_seguro.pdf" alt="Logo" width=350 />';
-        // echo '      </td>';
-        // echo '  </tr>';
-        // echo '</table>';
         }
     }
     ?>
